Initial Commit: intro to VQE added
This is commit 0 of this project. The project is supposed to host a
collection of educational open quantum information processing Jupyter
notebooks.

With this commit quantum_variational_algorithms/VA0_qubo_and_vqe.ipynb
has been added which is an introduction to Variational Quantum
Eigensolvers and a sample VQE for max-cut.