1-D Classical Coupled Harmonic Oscillators Pipeline Added
With this commit, a full tested pipeline for Hamiltonian simulation of
a 1-D classical system of N coupled harmonic oscillator have been
added.

The pipeline would first prepare a Pauli decomposition of the encoded
Hamiltonian matrix which would then be used in a Suzuki-trotter
decomposer to simulate the system evolution in time. The created
notebook contains the definition of all required functions and as a
final section, the analytical results are obtained to be compare with
the simulation results.

It should be noted that for the Suzuki-Trotter, order has been set to
4 and repetition to a dynamic value 2*i where i the the index of the
`evolution_coefficient` in the time interval considered.